requirements.txt
absl-py==2.1.0

altair==5.4.1

astunparse==1.6.3

attrs==24.2.0

blinker==1.8.2

cachetools==5.5.0

certifi==2024.8.30

cffi==1.17.1

charset-normalizer==3.3.2

click==8.1.7

colorama==0.4.6

contourpy==1.3.0

cycler==0.12.1

flatbuffers==24.3.25

fonttools==4.53.1

gast==0.6.0

gitdb==4.0.11

GitPython==3.1.43

google-pasta==0.2.0

grpcio==1.66.1

h5py==3.11.0

idna==3.8

jax==0.4.31

jaxlib==0.4.31

Jinja2==3.1.4

joblib==1.4.2

jsonschema==4.23.0

jsonschema-specifications==2023.12.1

kaleido==0.2.1

keras==3.5.0

kiwisolver==1.4.5

libclang==18.1.1

Markdown==3.7

markdown-it-py==3.0.0

MarkupSafe==2.1.5

matplotlib==3.9.2

mdurl==0.1.2

mediapipe==0.10.14

ml-dtypes==0.3.2

namex==0.0.8

narwhals==1.6.0

numpy==1.26.4

opencv-contrib-python==4.10.0.84

opencv-python==4.10.0.84

opt-einsum==3.3.0

optree==0.12.1

packaging==24.1

pandas==2.2.2

pillow==10.4.0

plotly==5.24.0

protobuf==4.25.4

pyarrow==17.0.0

pycparser==2.22

pydeck==0.9.1

Pygments==2.18.0

pyparsing==3.1.4

PyQt5==5.15.11

PyQt5-Qt5==5.15.2

PyQt5_sip==12.15.0

pyqtgraph==0.13.7

pyserial==3.5

python-dateutil==2.9.0.post0

pytz==2024.1

referencing==0.35.1

requests==2.32.3

rich==13.8.0

rpds-py==0.20.0

scikit-learn==1.5.2

scipy==1.14.1

seaborn==0.13.2

six==1.16.0

smmap==5.0.1

sounddevice==0.5.0

streamlit==1.38.0

tenacity==8.5.0

tensorboard==2.16.2

tensorboard-data-server==0.7.2

tensorflow==2.16.1

tensorflow-intel==2.16.1

tensorflow-io-gcs-filesystem==0.31.0

termcolor==2.4.0

threadpoolctl==3.5.0

toml==0.10.2

tornado==6.4.1

typing_extensions==4.12.2

tzdata==2024.1

urllib3==2.2.2

watchdog==4.0.2

Werkzeug==3.0.4

wrapt==1.16.0
